LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/ServiceWorkerGlobalScope/extendable-message-event-constructor.https-expected.txt


PASS ServiceWorkerGlobalScope: ExtendableMessageEvent Constructor 
PASS no initializer specified 
PASS `bubbles` is specified 
PASS `cancelable` is specified 
PASS `data` is specified 
PASS `origin` is specified 
PASS `lastEventId` is specified 
PASS `source` is specified 
FAIL `ports` is specified assert_throws: function "function () {
      createEvent({ get ports() { throw { name: 'Error' }; } }); }" threw object "TypeError: Value is not a sequence" ("TypeError") expected object "[object Object]" ("Error")
PASS all initial values are specified 
